# use buffer to reduce number of flush() .. (this also needed to use Writer instead of OutputStream/PrintStream)