Introducing FlushBufferedWriter between the AnsiPrintSteam and the underlying stdout to reduce the number of flush()

Also introducing similar FlushBufferedOutputSteam in deprecated AnsiOutputStream

Both classes still suffer from a problem of encoding, because they use java.io.OutputStream (to write byte,byte[]) instead of java.io.Writer (to write char,char[],String,CharSequence)

This is very infortunate that in java.io.PrintStream, the class is able to format both text oriented value and byte[] as it extends from java.io.OutputStream.
For write(int) and write(int[]) methods, I pretend that the bytes should NOT be interpreted as chars, and should NOT be filtered by ansi escape code. I have HOWEVER left the feature as it was.

For print() and println() values, the class PrintStream formats values as String then encode as byte[] using OutputStreamWriter ... then using jdk-private method flushBuffer() transfers byte[] to the underlying OutputStream without flushing.  This is the same logic to overwrite in the class jansi FilterPrintStream, using the same buffer mechanism.

Fragments of jdk code for flushBuffer() is copied here is this PR to allow both proper interpretation of char encoding, then avoid to many flush() on underlying stdout.