@Arnaud-Nauwynck 2019-06-01
I have split parts of the work done on my fork repo into several independent PRs,
as asked by @hboutemy 

This is the most important commit in terms of performance improvment.

There are still more PRs for me to submit if you want to also apply the refactoring to cleanup code duplication.

Please let me know if this work is worth doing it, if there is a chance you merge my pull-requests

This includes:
- change "class AnsiOutputStream extends OutputSteam" to "class AnsiOutputWriter extends Writer"
- remove code duplication between class AnsiOutputStream and AnsiPrintStream
     in particular, retain only 1 out of 2 classes, and remove all comments // expected diff with AnsiPrintStream.java
   .. so change FilterPrintSteam to simply delegate to a AnsiOutputWriter
- introduce class TerminalCommandProcessor to extract code specific to processing ansi escape commands  (currently mixed into AnsiOutputStream, and using override...)
   .. so change class inheritance by delegation...(Design Pattern best practices).
   .. so also refactor to class WindowsTerminalCommandProcessor extends TerminalCommandProcessor instead of extending both WindowsAnsiOutputStream and WindowsAnsiPrintStream
- introduce class TerminalType as a factory of TerminalCommandProcessor, for decorating a Writer and wrapping it into a AnsiOutputWriter