@Dunemaster 2019-08-03
@Arnaud-Nauwynck , what do you intend to do with the PR? Will you fix the conflicts?

Thank you