@Arnaud-Nauwynck 2019-09-27
Hi @Dunemaster 

Sorry for being so long after your comments on Aug 3.
I have merged the conflicts. Can you have a look?
Notice I still have other PRs to rebase after this commit... in particular for using Writer(char) instead of OutputStream(byte)
To test performance improvments, you need for example PR#149

Regards,
Arnaud