@Arnaud-Nauwynck 2019-09-28
I have run the benchmark-test provided in PR#149 with and without the optim for this PR

mvn package
 cd benchmark-test
java -cp "target/benchmark-test-1.19-SNAPSHOT.jar;../jansi/target/jansi-1.19-SNAPSHOT.jar" org.fusesource.jansi.BenchmarkMain

here are the results on a windows cmd.exe or powershell 

-------- with optim buffer-write, in windows powershell terminal 
loop Stdout Before 5x100 took 40ms
loop Jansi 5x100 took 199ms
loop Stdout 5x100 took 47ms
loop Jansi BigLines 5x100 took 148ms
loop Stdout BigLines 5x100 took 6ms
loop Jansi NoFlush 5x100 took 188ms
loop Stdout NoFlush 5x100 took 16ms
loop Overhead ansi + JansiPrintSteam 5x100 took 9ms
loop Overhead Raw JansiPrintSteam 5x100 took 0ms
count flush() ansi+JansiPrintStream 5x100 : 502
count flush() raw+JansiPrintStream 5x100 : 502

--------- witout the optim, in windows powershell terminal
lloop Stdout Before 5x100 took 33ms
loop Jansi 5x100 took 462ms
loop Stdout 5x100 took 338ms
loop Jansi BigLines 5x100 took 698ms
loop Stdout BigLines 5x100 took 656ms
loop Jansi NoFlush 5x100 took 527ms
loop Stdout NoFlush 5x100 took 412ms
loop Overhead ansi + JansiPrintSteam 5x100 took 15ms
loop Overhead Raw JansiPrintSteam 5x100 took 7ms
count flush() ansi+JansiPrintStream 5x100 : 2
count flush() raw+JansiPrintStream 5x100 : 2

--------- with the optim, and running with PASS_THROUGH mode on ConsoleEmu
loop Stdout Before 5x100 took 136ms
loop Jansi 5x100 took 422ms

interpretation:   doing some printlns in windows powershell terminal
without jansi :            40ms,
with jansi :                 462ms
with jansi optimized: 199ms
 => there is an overhead when using jansi... but the optimization of this PR make it x2.5 faster than without


The performance is dependent on the terminal screen size (reducing to the minimum make it faster..).
but jansi without the optim still have huge overhead (527ms) ... with the optim, the overhead is smaller (137ms)
This is the worst case for jansi without the optim, showing an overhead of ~x20 slower than println !! 

Summary: 
---------
The optimization of this PR make it x2.5 faster than without
Jansi still have a significant overhead compared to println ~x2 ... so maybe more optimisations to come.