@gnodet 2020-10-23
Fixed by https://github.com/fusesource/jansi/commit/57aa84d4120395922458f61f20e741198f051087